requirements.txt
python-vlc
requests
pytube
moviepy
YouTubeMusicAPI